all e2e tests in one place, clean up a bit
