+++ app/test/e2e/click-everything.e2e.ts
@@ -1,6 +1,6 @@
 import { test } from '@playwright/test'
 
-import { expectNotVisible, expectRowVisible, expectVisible } from 'app/test/e2e'
+import { expectNotVisible, expectRowVisible, expectVisible } from './utils'
 
 test.beforeEach(async ({ page }) => {
   await page.goto('/projects/mock-project')
