+++ app/test/e2e/disk-create.e2e.ts
@@ -1,4 +1,4 @@
-import { expectVisible, test } from 'app/test/e2e'
+import { expectVisible, test } from './utils'
 
 test.describe('Disk create', () => {
   test.beforeEach(async ({ page }) => {
