--- app/test/e2e/index.ts
@@ -1,2 +1,0 @@
-export * from '@playwright/test'
-export * from './utils'
