+++ app/test/e2e/project-access.e2e.ts
@@ -7,14 +7,10 @@ import {
   expectRowVisible,
   expectSimultaneous,
   expectVisible,
-} from 'app/test/e2e'
+} from './utils'
 
-test('Click through silo access page', async ({ page }) => {
-  await page.goto('/')
-
-  const table = page.locator('role=table')
-
-  // page is there, we see user 1 and 2 but not 3
+test('Click through project access page', async ({ page }) => {
+  await page.goto('/projects/mock-project')
   await page.click('role=link[name*="Access & IAM"]')
 
   // has to be before anything else is checked. ensures we've prefetched
@@ -25,32 +21,47 @@ test('Click through silo access page', async ({ page }) => {
     'role=cell[name="Hannah Arendt"]',
   ])
 
+  // page is there, we see user 1-3 but not 4
   await expectVisible(page, ['role=heading[name*="Access & IAM"]'])
+  const table = page.locator('table')
+  await expectRowVisible(table, {
+    Name: 'Hannah Arendt',
+    'Silo role': 'admin',
+    'Project role': '',
+  })
+  await expectRowVisible(table, {
+    Name: 'Jacob Klein',
+    'Silo role': '',
+    'Project role': 'collaborator',
+  })
   await expectRowVisible(table, {
     // no space because expectRowVisible uses textContent, not accessible name
     Name: 'real-estate-devsGroup',
     'Silo role': 'admin',
   })
   await expectRowVisible(table, {
-    Name: 'Hannah Arendt',
-    'Silo role': 'admin',
+    // no space because expectRowVisible uses textContent, not accessible name
+    Name: 'kernel-devsGroup',
+    'Silo role': '',
+    'Project role': 'viewer',
   })
+
   await expectNotVisible(page, [`role=cell[name="${user4.display_name}"]`])
 
-  // Add user 2 as collab
+  // Add user 4 as collab
   await page.click('role=button[name="Add user or group"]')
   await expectVisible(page, ['role=heading[name*="Add user or group"]'])
 
   await page.click('role=button[name="User or group"]')
-  // only users not already on the org should be visible
-  await expectNotVisible(page, ['role=option[name="Hannah Arendt"]'])
+  // only users not already on the project should be visible
+  await expectNotVisible(page, ['role=option[name="Jacob Klein"]'])
   await expectVisible(page, [
+    'role=option[name="Hannah Arendt"]',
     'role=option[name="Hans Jonas"]',
-    'role=option[name="Jacob Klein"]',
     'role=option[name="Simone de Beauvoir"]',
   ])
 
-  await page.click('role=option[name="Jacob Klein"]')
+  await page.click('role=option[name="Simone de Beauvoir"]')
 
   await page.click('role=button[name="Role"]')
   await expectVisible(page, [
@@ -62,15 +73,15 @@ test('Click through silo access page', async ({ page }) => {
   await page.click('role=option[name="Collaborator"]')
   await page.click('role=button[name="Assign role"]')
 
-  // User 3 shows up in the table
+  // User 4 shows up in the table
   await expectRowVisible(table, {
-    Name: 'Jacob Klein',
-    'Silo role': 'collaborator',
+    Name: 'Simone de Beauvoir',
+    'Project role': 'collaborator',
   })
 
-  // now change user 3's role from collab to viewer
+  // now change user 4 role from collab to viewer
   await page
-    .locator('role=row', { hasText: user3.display_name })
+    .locator('role=row', { hasText: user4.display_name })
     .locator('role=button[name="Row actions"]')
     .click()
   await page.click('role=menuitem[name="Change role"]')
@@ -82,9 +93,10 @@ test('Click through silo access page', async ({ page }) => {
   await page.click('role=option[name="Viewer"]')
   await page.click('role=button[name="Update role"]')
 
-  await expectRowVisible(table, { Name: user3.display_name, 'Silo role': 'viewer' })
+  await expectRowVisible(table, { Name: user4.display_name, 'Project role': 'viewer' })
 
-  // now delete user 3
+  // now delete user 3. has to be 3 or 4 because they're the only ones that come
+  // from the project policy
   await expectVisible(page, [`role=cell[name="${user3.display_name}"]`])
   await page
     .locator('role=row', { hasText: user3.display_name })
@@ -92,4 +104,17 @@ test('Click through silo access page', async ({ page }) => {
     .click()
   await page.click('role=menuitem[name="Delete"]')
   await expectNotVisible(page, [`role=cell[name="${user3.display_name}"]`])
+
+  // now add a project role to user 1, who currently only has silo role
+  await page.click('role=button[name="Add user or group"]')
+  await page.click('role=button[name="User or group"]')
+  await page.click('role=option[name="Hannah Arendt"]')
+  await page.click('role=button[name="Role"]')
+  await page.click('role=option[name="Viewer"]')
+  await page.click('role=button[name="Assign role"]')
+  await expectRowVisible(table, {
+    Name: 'Hannah Arendt',
+    'Silo role': 'admin',
+    'Project role': 'viewer',
+  })
 })
