+++ app/test/e2e/top-bar.e2e.ts
@@ -1,10 +1,9 @@
 import { test } from '@playwright/test'
 
-import { expectSimultaneous } from 'app/test/e2e'
-import { pb } from 'app/util/path-builder'
+import { expectSimultaneous } from './utils'
 
 test('Silo/system picker does not pop in', async ({ page }) => {
-  await page.goto(pb.projects())
+  await page.goto('/projects')
 
   // make sure the system policy call is prefetched properly so that the
   // silo/system picker doesn't pop in. if this turns out to be flaky, just
