+++ playwright.config.ts
@@ -5,7 +5,7 @@ import { devices } from '@playwright/test'
  * See https://playwright.dev/docs/test-configuration.
  */
 const config: PlaywrightTestConfig = {
-  testDir: './app',
+  testDir: './app/test/e2e',
   testMatch: /\.e2e\.ts/,
   /* Fail the build on CI if you accidentally left test.only in the source code. */
   forbidOnly: !!process.env.CI,
